Create solution1.js
This code is using the jQuery library to create a simple web page that allows the user to search for gifs using the Giphy API.

The first two lines of code select the HTML elements with the ids of "gif-area" and "search" and assigns them to the variables $gifArea and $searchInput respectively.
The function "addGif" takes in a response from the Giphy API and creates a new div element with the class "col-md-4 col-12 mb-4" and an image element with the src set to the url of a random gif from the API response. It then appends the image element to the div and the div to the $gifArea element.
The next block of code listens for a submit event on the form element. On submit, it prevents the default behavior, gets the value of the search input, clears the search input and makes an AJAX call to the Giphy API using the axios library, passing the searchTerm as a query parameter. Once the API call is successful, it passes the response data to the addGif function.
The last block of code listens for a click event on an element with an id of "remove" and, when clicked, it empties the $gifArea element, effectively removing all gifs currently displayed.